Update README.md
Kilt Outfit is a family-run business-based company with a strong presence in the online market. As shown from the name, we specialize in custom-made jackets and tartan fabric products. We have been working in this field for the last several years. We have a dedicated team of 40  professional people who are passionate about bringing unique and exclusive designs. With years of experience in the field, we have now become trendsetters. We dedicate a lot of time and effort to new designs. Kilt Outfit has a huge variety of Scottish jackets that are unique and you have never seen as before. We understand our customers and their requirements about what they want in Custom made products. We believe that a custom-made product gives a more satisfactory look and perfect fitting than a readymade product. We feel proud to offer a custom-made facility to our customers. These custom-made products take 3-4 weeks in manufacturing and if you need to get an order early then you should go for a rush order. We have kept prices affordable and customer-friendly. All our Jackets and other products are handmade. You are very welcome to pop in and take a look at our products.